requirements.txt
numpy
pandas
scikit-learn
joblib
huggingface-hub
python-dotenv
